+++ Src/NanoRoute/Public/Extensions/NanoRouteJsonExtensions.cs
@@ -606,9 +606,7 @@ namespace NanoRoute
                         }
                         catch (HttpRequestException ex)
                         {
-                            context.Request.Properties.TryGetValue(Router.TraceIdName, out object? traceId);
-
-                            ErrorDetails errorDetails = ex.GetErrorDetails(config.PopulateErrorInfo, traceId as string);
+                            ErrorDetails errorDetails = ex.GetErrorDetails(config.PopulateErrorInfo, context.Request.TraceId);
 
                             return HttpResponseMessage.Json
                             (